# payup-tm/apidoc.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/Architecton/PayUp/1.0.0"
  } ],
  "info" : {
    "description" : "This is the API for the PayUp loan management application.\n",
    "version" : "1.0.0",
    "title" : "PayUp"
  },
...
  "paths" : {
    "/users" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get all users in the database",
        "operationId" : "userGetAll",
        "responses" : {
          "404" : {
            "description" : "Cannot find users"
          },
          "500" : {
            "description" : "Error while getting users"
          },
          "200" : {
            "description" : "Users were successfully retrieved",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Add a new user to the database",
        "operationId" : "userAddNew",
        "responses" : {
          "400" : {
            "description" : "Invalid user parameters"
          },
          "500" : {
            "description" : "Error creating new user or sending confirmation e-mail"
          },
          "201" : {
            "description" : "User successfully created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string"
                  },
                  "surname" : {
                    "type" : "string"
                  },
                  "_id" : {
                    "type" : "string",
                    "description" : "username - primary key"
                  },
                  "password" : {
                    "type" : "string"
                  },
                  "email" : {
                    "type" : "string"
                  },
                  "gender" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{idUser}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Find user by ID",
        "description" : "Returns a single user matching the ID",
        "operationId" : "getUserByID",
        "parameters" : [ {
          "name" : "idUser",
          "in" : "path",
          "description" : "user's ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "User with specified ID not found"
          },
          "500" : {
            "description" : "Error retrieving user"
          },
          "200" : {
            "description" : "User successfully retrieved",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Users" ],
        "summary" : "Delete a user from the database",
        "operationId" : "userDelete",
        "parameters" : [ {
          "name" : "idUser",
          "in" : "path",
          "description" : "User id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "User with specified ID not found"
          },
          "204" : {
            "description" : "User successfully deleted"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "User" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "_id" : {
            "type" : "string",
            "description" : "user's username serves as the id"
          },
          "password" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "gender" : {
            "type" : "string"
          },
          "dateJoined" : {
            "type" : "object",
            "description" : "date - is set to current date on creation"
          },
          "status" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "User Status 0 ~ activated_no; 1 ~ activated_yes; 2 ~ terminated"
          },
          "defaultCurrency" : {
            "type" : "string"
          },
          "nightmode" : {
            "type" : "boolean"
          },
          "loans" : {
            "type" : "object",
            "description" : "list of user's loans"
          },
          "contacts" : {
            "type" : "object",
            "description" : "list of user's contacts"
          }
        }
      },
...
    }
  }
}